Pass the jigna qt tests now.
Need to fix:
 - test_html_widget requires user to close one of the windows first.
 - file loader is not working properly.
